requirements.txt
backend==0.2.4.1
fastapi==0.118.0
langchain_chroma==0.2.6
langchain_community==0.3.30
langchain_experimental==0.3.4
langchain_openai==0.3.34
openai==2.1.0
pydantic==2.11.10
python-dotenv==1.1.1
starlette==0.48.0
uvicorn==0.37.0
python-multipart
pymupdf

azure-core~=1.38.0
oci-cli~=3.74.0